starbases can now be attacked and destroyed.

+++ Ship.gd
@@ -66,30 +66,29 @@ func _input_event(viewport, event, shape_idx):
 	# fighting costs 2 fuel
 	elif event.type == InputEvent.MOUSE_BUTTON and event.button_index == BUTTON_LEFT and event.is_pressed() and get_tree().get_root().get_node("Game").players[current_player - 1].fuel > 2:
 		# compile fleets in combat
-		var enemy_fleet = [] # the current player's fleet
-		var enemy_ammo = 0
-		var allied_fleet = [] # this ship (the target)'s fleet
-		var allied_ammo = 0
-		for s in get_parent().get_ships():
-			if s.owner_id == current_player:
-				enemy_fleet.append(s)
-			elif s.owner_id == owner_id:
-				allied_fleet.append(s)
-		# calculate casualties
-		var allied_casualties = max(1, int(round(enemy_fleet.size() / 2)))
-		var enemy_casualties = max(1, int(round(allied_fleet.size() / 2)))
-		if planet.starbases.size() > 0:
-			if planet.starbases[0].owner_id == current_player:
-				allied_casualties += 2
-			elif planet.starbases[0].owner_id == owner_id:
-				enemy_casualties += 2
-		# resolve combat, destroy ships
-		for i in range(allied_casualties):
-			if allied_fleet.size() > 0:
-				planet.destroy_ship(allied_fleet[0])
-				allied_fleet.remove(0)
-		for i in range(enemy_casualties):
-			if enemy_fleet.size() > 0:
-				planet.destroy_ship(enemy_fleet[0])
-				enemy_fleet.remove(0)
+		if get_parent().get_selected_ships_by_owner(get_tree().get_root().get_node("Game").players[current_player - 1]).size() > 0:
+			var enemy_fleet = [] # the current player's fleet
+			var allied_fleet = [] # this ship (the target)'s fleet
+			for s in get_parent().get_ships():
+				if s.owner_id == current_player:
+					enemy_fleet.append(s)
+				elif s.owner_id == owner_id:
+					allied_fleet.append(s)
+			# calculate casualties
+			var allied_casualties = max(1, int(round(enemy_fleet.size() / 2)))
+			var enemy_casualties = max(1, int(round(allied_fleet.size() / 2)))
+			if planet.starbases.size() > 0:
+				if planet.starbases[0].owner_id == current_player:
+					allied_casualties += 2
+				elif planet.starbases[0].owner_id == owner_id:
+					enemy_casualties += 2
+			# resolve combat, destroy ships
+			for i in range(allied_casualties):
+				if allied_fleet.size() > 0:
+					planet.destroy_ship(allied_fleet[0])
+					allied_fleet.remove(0)
+			for i in range(enemy_casualties):
+				if enemy_fleet.size() > 0:
+					planet.destroy_ship(enemy_fleet[0])
+					enemy_fleet.remove(0)
 	planet.update_ship_display()
